+++ experiments/winui3-main-window/winui-app/pch.h
@@ -10,6 +10,7 @@
 #undef GetCurrentTime
 
 #include <winrt/Windows.ApplicationModel.h>
+#include <winrt/Windows.ApplicationModel.Activation.h>
 #include <winrt/Windows.ApplicationModel.DataTransfer.h>
 #include <winrt/Windows.Data.Json.h>
 #include <winrt/Windows.Foundation.h>